# solution/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления данными клиентов, рекламодателей, рекламными кампаниями, показом объявлений, статистикой и управлением \"текущим днём\" в системе.",
        "title": "PROD Backend 2025 Advertising Platform API",
        "contact": {},
        "license": {
            "name": "GPL 3.0"
        },
        "version": "1.0"
    },
    "paths": {
        "/ads/{adId}/click": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Ads"
                ],
                "summary": "Фиксация перехода по рекламному объявлению",
                "parameters": [
                    {
                        "description": "click",
                        "name": "click",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Click"
                        }
                    },
                    {
                        "type": "string",
                        "description": "UUID рекламного объявления (идентификатор кампании), по которому совершен клик",
                        "name": "adId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/advertisers/bulk": {
            "post": {
                "description": "Создает или обновляет рекламодателей",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advertisers"
                ],
                "summary": "Создание/обновление рекламодателей",
                "parameters": [
                    {
                        "description": "CampaignRequest",
                        "name": "CreateAdvertisers",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Advertiser"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Advertiser"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/advertisers/{advertiserId}": {
            "get": {
                "description": "Возвращает информацию о рекламодателе по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advertisers"
                ],
                "summary": "Получение рекламодателя по ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Advertiser"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/ml-scores": {
            "post": {
                "description": "Добавляет или обновляет ML скор для указанной пары клиент-рекламодатель",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advertisers"
                ],
                "summary": "Добавление или обновление ML скора",
                "parameters": [
                    {
                        "description": "MLScore",
                        "name": "MLScore",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.MLScore"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Advertiser": {
            "type": "object",
            "properties": {
                "advertiser_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "domain.Click": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.MLScore": {
            "type": "object",
            "properties": {
                "advertiser_id": {
                    "type": "string"
                },
                "client_id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
...
    }
}